pull out createNewRoute
